+++ CHANGELOG.md
@@ -4,7 +4,18 @@ Mapbox welcomes participation and contributions from everyone.
 
 ### main
 
+- Added `StepIntersection#formOfWay` property which provides a list representing the "form of way" values for all roads at the step intersection. [#1611](https://github.com/mapbox/mapbox-java/pull/1611)
+- Added `StepIntersection#geometries` property which provides a list representing the geometry of each road at the step intersection. [#1611](https://github.com/mapbox/mapbox-java/pull/1611)
+- Added `StepIntersection#access` property which provides a list representing the access type for each road at the step intersection. [#1611](https://github.com/mapbox/mapbox-java/pull/1611)
+- Added `StepIntersection#elevated` property which provides a list indicating whether each road at the step intersection is elevated. [#1611](https://github.com/mapbox/mapbox-java/pull/1611)
+- Added `StepIntersection#bridge` property which provides a list indicating whether each road at the step intersection is a bridge. [#1611](https://github.com/mapbox/mapbox-java/pull/1611)
+- Added a new voice unit value: `DirectionsCriteria#BRITISH_IMPERIAL`. This value is now included in `DirectionsCriteria#VoiceUnitCriteria`. [#1611](https://github.com/mapbox/mapbox-java/pull/1611)
+
+
+### v7.5.0 - July 31, 2025
+
 - Updated `auto-value-gson` to version [0.0.3](https://github.com/mapbox/auto-value-gson/releases/tag/mapbox-v0.0.3) and `gson` to version 2.13.1. [#1615](https://github.com/mapbox/mapbox-java/pull/1615)
+
 
 ### v7.4.0 - April 11, 2025